change code to single precision

+++ distance_est.cc
@@ -22,11 +22,11 @@
 #include "mandelbox.h"
 
 extern MandelBoxParams mandelBox_params;
-extern double MandelBoxDE(const vec3 &pos, const MandelBoxParams &mPar, double c1, double c2);
-extern double MandelBulbDistanceEstimator(const vec3 &p0, const MandelBoxParams &params);
+extern float MandelBoxDE(const vec3 &pos, const MandelBoxParams &mPar, float c1, float c2);
+extern float MandelBulbDistanceEstimator(const vec3 &p0, const MandelBoxParams &params);
 
 //Distance Estimator Field Selector
-double DE(const vec3 &p)
+float DE(const vec3 &p)
 {
   //double c1 = fabs(mandelBox_params.scale - 1.0);
   //double c2 = pow( fabs(mandelBox_params.scale), 1 - mandelBox_params.num_iter);
