+++ mandelbulb_dist_est.cc
@@ -29,40 +29,40 @@
 
 
 
-double MandelBulbDistanceEstimator(const vec3 &p0, const MandelBoxParams &params)
+float MandelBulbDistanceEstimator(const vec3 &p0, const MandelBoxParams &params)
 {
   vec3 z;
   z = p0;
   
-  double dr = 1.0;
-  double r = 0.0;
+  float dr = 1.0;
+  float r = 0.0;
 
-  double Bailout = params.rMin;
-  double Power = params.rFixed;
+  float Bailout = params.rMin;
+  float Power = params.rFixed;
 
   for (int i=0; i < params.num_iter; i++) 
     {
       MAGNITUDE(r,z);
       if(r > Bailout) break; 
 
-      double theta = acos(z.z/r);
-      double phi   = atan2(z.y, z.x);
-      dr = pow(r, Power - 1.0) * Power * dr + 1.0;
+      float theta = acosf(z.z/r);
+      float phi   = atan2f(z.y, z.x);
+      dr = powf(r, Power - 1.0) * Power * dr + 1.0;
 
-      double zr = pow(r, Power);
+      float zr = powf(r, Power);
       theta     = theta * Power;
       phi       = phi * Power;
 
-      z.x = zr*sin(theta)*cos(phi);
-      z.y = zr*sin(phi)*sin(theta);
-      z.z = zr*cos(theta);
+      z.x = zr*sinf(theta)*cosf(phi);
+      z.y = zr*sinf(phi)*sinf(theta);
+      z.z = zr*cosf(theta);
 
       z.x = z.x + p0.x;
       z.y = z.y + p0.y;
       z.z = z.z + p0.z;
     }
 
-  return 0.5*log(r)*r/dr;
+  return 0.5*logf(r)*r/dr;
 }
